branch

+++ app/Http/Controllers/AuthUserController.php
@@ -2,51 +2,51 @@
 
 namespace App\Http\Controllers;
 
-use Str;
-use Auth;
-use App\Models\User;
-use Illuminate\Http\Request;
 use App\Http\Requests\AuthUserRequest;
+use App\Models\User;
+use Auth;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
 
 class AuthUserController extends Controller
 {
-    public function login (AuthUserRequest $request){
+    public function login(AuthUserRequest $request): JsonResponse
+    {
         $validatedData = $request->validated();
-
+        if (! $validatedData) {
+            return response()->json([
+                'message' => 'invalid data',
+            ]);
+        }
         if (Auth::attempt(['email' => $validatedData['email'],
-         'password' => $validatedData['password']])){
-
+            'password' => $validatedData['password']])) {
             $user = Auth::User();
             $token = $user->createToken('auth_token')->plainTextToken;
 
             return response()->json([
                 'message' => 'Login realized with sucess',
                 'user' => $user,
-                'token' => $token
-            ], 200);
-         }
+                'token' => $token,
+            ]);
+        }
 
-
-         return response()->json([
-            'message' => 'Invalid information!'
-         ]);
+        return response()->json([
+            'message' => 'Invalid data!',
+        ], 401);
     }
 
-
-
-    public function register (){
-
-        return response()->json(route('store'));
-    }
-
-    public function logout (Request $request){
-            $user = User::find(Auth::id());
-            $user->tokens()->delete();
+    public function logout(Request $request): JsonResponse
+    {
+        $user = User::find(Auth::id());
+        if (! $user) {
+            return response()->json([
+                'message' => 'user not found',
+            ], 404);
+        }
+        $user->tokens()->delete();
 
         return response()->json([
             'message' => 'logout realized with sucess!',
-
         ]);
     }
 }
-
